requirements.txt
opencv-python>=4.8
numpy>=1.24
ultralytics>=8.1
pyyaml>=6.0
tqdm>=4.65
